+++ packages/twopoint5d/src/stage/public-api.ts
@@ -1,7 +1,9 @@
 export * from './Canvas2DStage.js';
 export * from './fitIntoRectangle.js';
 export type * from './IProjection.js';
+export type * from './IRenderable.js';
 export type * from './IStage.js';
+export type * from './IStageRendererHost.js';
 export * from './OrthographicProjection.js';
 export * from './ParallaxProjection.js';
 export * from './ProjectionPlane.js';